Admin routing condition added
