--- app/admin/roles/page.tsx
@@ -1,12 +1,0 @@
-import React from "react";
-import UserRoleManagement from "./RoleManagement";
-
-const RoleManagementPage = () => {
-  return (
-    <div>
-      <UserRoleManagement />
-    </div>
-  );
-};
-
-export default RoleManagementPage;
