+++ app/api/auth/[...nextauth]/route.tsx
@@ -1,17 +1,7 @@
 import NextAuth from "next-auth";
 import CredentialsProvider from "next-auth/providers/credentials";
 
-interface User {
-  email: string;
-  firstName: string;
-}
-interface Token {
-  email: string;
-  firstName: string;
-  accessToken: string;
-}
-
-const handler = NextAuth({
+export const handler = NextAuth({
   providers: [
     CredentialsProvider({
       name: "Credentials",
@@ -20,9 +10,6 @@ const handler = NextAuth({
         password: { label: "Password", type: "password" },
       },
       async authorize(credentials) {
-        // console.log(credentials);
-        // console.log(`${process.env.NEXT_PUBLIC_BACKEND_DOMAIN}/user/signin`);
-
         const authResponse = await fetch(
           `${process.env.NEXT_PUBLIC_BACKEND_DOMAIN}/user/signin`,
           {
@@ -36,19 +23,12 @@ const handler = NextAuth({
             }),
           },
         );
-        // console.log("Response",authResponse)
 
         if (!authResponse.ok) {
           return null;
         }
 
         const user = await authResponse.json();
-
-        console.log("userdetails", user);
-        const Role = user.data.role;
-        console.log(Role);
-
-        // console.log("Session  details:", user.data.first_name,user.data.email,user.data.role_name,user.data.file_name,user.data.id);
 
         return {
           email: user.data.email,
@@ -62,25 +42,15 @@ const handler = NextAuth({
   ],
 
   callbacks: {
-    async session({ session, user, token }) {
-      // console.log(user.role_name);
-      //  session.user={...session.user,role_name:"Admin"}
-      console.log("user", user);
-      session.accessToken = token.accessToken;
-      session.user.id = token.id;
-      session.role = user.data.role_name;
-      console.log(session);
-      console.log("session:", session.user);
-      // console.log("user",user)
-      return session;
+    async jwt({ token, user }) {
+      if (user) token.role = user.role;
+      return token;
     },
 
-    async jwt({ token, account, profile }) {
-      if (account) {
-        token.accessToken = "testAccessToken";
-        token.id = "profile.id";
-      }
-      return token;
+    async session({ session, token }) {
+      if (session?.user) session.user.role = token.role;
+
+      return session;
     },
   },
 });
