+++ app/homepage/layout.tsx
@@ -1,5 +1,5 @@
 // components/Layout.js
-import React from "react";
+// import React from "react";
 import { HomePage } from "../components/HomePage";
 import FooterPart from "../components/FooterPart";
 import VideoPart from "./VideoPart";
