+++ app/middleware.tsx
@@ -1,0 +1,3 @@
+import { withAuth } from "next-auth/middleware";
+
+export const config = { matcher: ["/admin"] };
